Stage1-fixed

+++ src/models/layers/Residual.py
@@ -1,3 +1,4 @@
+import torch
 import torch.nn as nn
 
 class Residual(nn.Module):
@@ -7,11 +8,11 @@ class Residual(nn.Module):
     self.numOut = numOut
     self.bn = nn.BatchNorm2d(self.numIn)
     self.relu = nn.ReLU(inplace = True)
-    self.conv1 = nn.Conv2d(self.numIn, self.numOut / 2, bias = True, kernel_size = 1)
-    self.bn1 = nn.BatchNorm2d(self.numOut / 2)
-    self.conv2 = nn.Conv2d(self.numOut / 2, self.numOut / 2, bias = True, kernel_size = 3, stride = 1, padding = 1)
-    self.bn2 = nn.BatchNorm2d(self.numOut / 2)
-    self.conv3 = nn.Conv2d(self.numOut / 2, self.numOut, bias = True, kernel_size = 1)
+    self.conv1 = nn.Conv2d(self.numIn, int(self.numOut/2), bias = True, kernel_size = 1)
+    self.bn1 = nn.BatchNorm2d(int(self.numOut/2))
+    self.conv2 = nn.Conv2d(int(self.numOut/2), int(self.numOut/2), bias = True, kernel_size = 3, stride = 1, padding = 1)
+    self.bn2 = nn.BatchNorm2d(int(self.numOut/2))
+    self.conv3 = nn.Conv2d(int(self.numOut/2), self.numOut, bias = True, kernel_size = 1)
     
     if self.numIn != self.numOut:
       self.conv4 = nn.Conv2d(self.numIn, self.numOut, bias = True, kernel_size = 1) 
@@ -32,4 +33,3 @@ class Residual(nn.Module):
       residual = self.conv4(x)
     
     return out + residual
-
